Apartado Materiales finalizado

+++ app/src/main/java/co/dolmen/sid/entidad/MovimientoArticulo.java
@@ -31,7 +31,7 @@ public class MovimientoArticulo implements Serializable {
         return id_tipo_stock;
     }
 
-    public double getCantidad() {
+    public float getCantidad() {
         return cantidad;
     }
